+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-*
+* Made part of the frontend code conditional on `NODE_ENV = "development"`
 
 ### Fixed